+++ src/components/login/googlelogin/Googlelogin.tsx
@@ -1,36 +1,15 @@
 'use client';
 import Image from 'next/image';
 import googlelogo from '@/assets/login/google.svg';
-import { useSearchParams } from 'next/navigation';
 import useLoginStore from '@/store/useLoginStore';
-import { useEffect } from 'react';
-import { useRouter } from 'next/navigation';
 
 export default function GoogleLogin() {
-  const router = useRouter();
-  const params = useSearchParams();
-
-  const accessToken = params.get('accessToken') || '';
-  const refreshToken = params.get('refreshToken');
-
-  //로그인 성공 여부
   const isLoggined = useLoginStore((state) => state.isLoggined);
   const setIsLoggined = useLoginStore((state) => state.setIsLoggined);
-
-  useEffect(() => {
-    if (accessToken && refreshToken) {
-      localStorage.setItem('accessToken', accessToken);
-      localStorage.setItem('refreshToken', refreshToken);
-      setIsLoggined(true);
-      router.push('/login');
-    }
-  }, [params]);
 
   // 로그아웃은 실제 /login 페이지에는 없지만, 로그인 구현을 위해 임시 생성
   const handleLogout = () => {
     setIsLoggined(false);
-
-    // todo : 로그아웃 처리 로직 구현하기
   };
 
   return (
